feat: implement AI-powered worker matchmaking for gigs with location filtering and fallback
